[Move #359] mafiaSocket" 폴더 이름 "mafia"로 변경
