+++ utils/mafia/getPlayerJob.ts
