+++ utils/mafia/getPlayersNumber.ts
